+++ ci/travis/audit-script.sh
@@ -1,4 +1,9 @@
 #!/usr/bin/env bash
 set -ex
 
+# Audit main workspace.
+cargo audit
+
+# Audit replidev crate.
+cd devtools/replidev
 cargo audit