+++ devtools/replidev/src/command/deps.rs
@@ -1,13 +1,18 @@
+use std::collections::BTreeMap;
 use std::fs::File;
 
 use failure::ResultExt;
+use prettytable::cell;
+use prettytable::row;
+use serde::Deserialize;
 use structopt::StructOpt;
 
 use crate::conf::Conf;
-use crate::conf::Project;
 use crate::podman::Pod;
 use crate::ErrorKind;
 use crate::Result;
+
+const PODMAN_DEF_PATH: &str = "devtools/podman";
 
 /// Manage Replicante Core dependencies.
 #[derive(Debug, StructOpt)]
@@ -19,6 +24,10 @@ pub enum CliOpt {
     /// Run the initialise command of each container in the given dependencies pods.
     #[structopt(name = "initialise")]
     Initialise(PodOpt),
+
+    /// List running and available dependencies pods.
+    #[structopt(name = "list")]
+    List,
 
     /// Stop and start supported dependencies pods.
     #[structopt(name = "restart")]
@@ -50,27 +59,34 @@ pub struct PodOpt {
     pods: Vec<String>,
 }
 
+#[derive(Debug, Deserialize)]
+struct PodPsStatus {
+    pub id: String,
+    pub status: String,
+}
+
 /// Manage Replicante Core dependencies.
 pub fn run(args: CliOpt, conf: Conf) -> Result<bool> {
-    if conf.project != Project::Core {
+    if !conf.project.allow_deps() {
         let error = ErrorKind::invalid_project(conf.project, "replidev deps");
         return Err(error.into());
     }
     match args {
         CliOpt::Clean(args) => clean(&args, &conf),
         CliOpt::Initialise(args) => initialise(&args, &conf),
+        CliOpt::List => list(&conf),
         CliOpt::Restart(args) => restart(&args, &conf),
         CliOpt::Start(args) => start(&args, &conf),
         CliOpt::Stop(args) => stop(&args, &conf),
     }
 }
 
-fn clean(args: &CleanOpt, _: &Conf) -> Result<bool> {
+fn clean(args: &CleanOpt, conf: &Conf) -> Result<bool> {
     for pod_name in &args.pod_opt.pods {
         let data = format!("./devtools/data/{}", pod_name);
         println!("--> Clean data for {} pod (from {})", pod_name, data);
         if args.confirm {
-            crate::podman::unshare(vec!["rm", "-r", &data])?;
+            crate::podman::unshare(conf, vec!["rm", "-r", &data])?;
         } else {
             println!("Skipping: you must --confirm deleting data");
         }
@@ -78,7 +94,7 @@ fn clean(args: &CleanOpt, _: &Conf) -> Result<bool> {
     Ok(true)
 }
 
-fn initialise(args: &PodOpt, _: &Conf) -> Result<bool> {
+fn initialise(args: &PodOpt, conf: &Conf) -> Result<bool> {
     for pod_name in &args.pods {
         let pod = pod_definition(pod_name)?;
         for container in pod.containers {
@@ -88,16 +104,91 @@ fn initialise(args: &PodOpt, _: &Conf) -> Result<bool> {
                     "--> Initialise {}/{} from {}",
                     pod_name, container.name, name
                 );
-                crate::podman::exec(&name, command)?;
+                crate::podman::exec(conf, &name, command)?;
             }
         }
     }
     Ok(true)
 }
 
+/// List running and available dependencies pod.
+///
+/// Example output?
+///   NAME      STATUS  POD ID  DEFINITION
+///   essential Running abc-123 $PODMAN_DEF_PATH/essential.yaml
+///   uis       -       -       $PODMAN_DEF_PATH/uis.yaml
+///   legacy    Running def-456 -
+fn list(conf: &Conf) -> Result<bool> {
+    // Find running dependencies pods.
+    let pods = crate::podman::pod_ps(
+        conf,
+        r#"{{ .Name }}: {id: "{{ .ID }}", status: "{{ .Status }}"}"#,
+        vec![
+            "label=io.replicante.dev.role=deps",
+            &format!("label=io.replicante.dev.project={}", conf.project),
+        ],
+    )?;
+    let pods: BTreeMap<String, PodPsStatus> =
+        serde_yaml::from_slice(&pods).expect("formatted podman pod ps output to parse");
+    let pods: BTreeMap<String, PodPsStatus> = pods.into_iter().map(|(key, value)| {
+        let key = key.trim_start_matches("replideps-").to_string();
+        (key, value)
+    })
+    .collect();
+
+    // Find available replideps definitions.
+    let mut available = Vec::new();
+    let dir = std::fs::read_dir(PODMAN_DEF_PATH)
+        .with_context(|_| ErrorKind::fs_error("failed to list available pods"))?;
+    for file in dir {
+        let file = file.with_context(|_| ErrorKind::fs_error("failed to list available pods"))?;
+        let name = file
+            .file_name()
+            .into_string()
+            .map_err(|_| ErrorKind::fs_error("failed to decode file name"))?;
+        if !name.ends_with(".yaml") {
+            continue;
+        }
+        let name = name.trim_end_matches(".yaml").to_string();
+        if pods.contains_key(&name) {
+            continue;
+        }
+        available.push(name);
+    }
+    available.sort();
+
+    // Print a table with the pods and info.
+    let mut table = prettytable::Table::new();
+    table.add_row(row!["NAME", "STATUS", "POD ID", "DEFINITION"]);
+    for pod in pods {
+        let name = pod.0;
+        let info = pod.1;
+        let path = format!("{}/{}.yaml", PODMAN_DEF_PATH, name);
+        let def = std::path::Path::new(&path);
+        let def = if def.exists() {
+            path
+        } else {
+            "-".to_string()
+        };
+        table.add_row(row![name, info.status, info.id, def]);
+    }
+    for name in available {
+        let name = format!("{}/{}.yaml", PODMAN_DEF_PATH, name);
+        table.add_row(row!["-", "-", "-", name]);
+    }
+
+    let format = prettytable::format::FormatBuilder::new()
+        .column_separator(' ')
+        .padding(0, 3)
+        .build();
+    table.set_format(format);
+    table.printstd();
+    Ok(true)
+}
+
 /// Helper function to load and decode a pod definition file.
 fn pod_definition(name: &str) -> Result<Pod> {
-    let definition = format!("devtools/podman/{}.yaml", name);
+    let definition = format!("{}/{}.yaml", PODMAN_DEF_PATH, name);
     let pod = File::open(definition).with_context(|_| ErrorKind::pod_not_found(name))?;
     let pod = serde_yaml::from_reader(pod).with_context(|_| ErrorKind::invalid_pod(name))?;
     Ok(pod)
@@ -113,6 +204,7 @@ fn start(args: &PodOpt, conf: &Conf) -> Result<bool> {
         let pod = pod_definition(pod_name)?;
         let variables = crate::podman::Variables::new(pod_name);
         crate::podman::pod_start(
+            conf,
             pod,
             format!("replideps-{}", pod_name),
             &conf.project,
@@ -122,9 +214,9 @@ fn start(args: &PodOpt, conf: &Conf) -> Result<bool> {
     Ok(true)
 }
 
-fn stop(args: &PodOpt, _: &Conf) -> Result<bool> {
+fn stop(args: &PodOpt, conf: &Conf) -> Result<bool> {
     for pod_name in args.pods.iter().rev() {
-        if crate::podman::pod_stop(format!("replideps-{}", pod_name)).is_err() {
+        if crate::podman::pod_stop(conf, format!("replideps-{}", pod_name)).is_err() {
             println!(
                 "--> Failed to stop {} pod, assuming it was not running",
                 pod_name