+++ devtools/replidev/src/lib.rs
@@ -5,6 +5,8 @@ mod conf;
 mod error;
 mod podman;
 
+use self::conf::Conf;
+
 pub use self::error::Error;
 pub use self::error::ErrorKind;
 pub use self::error::Result;