+++ devtools/replidev/src/podman/exec_cmd.rs
@@ -2,12 +2,13 @@ use std::process::Command;
 
 use failure::ResultExt;
 
+use crate::Conf;
 use crate::ErrorKind;
 use crate::Result;
 
 /// Execute a command in a container.
-pub fn exec(name: &str, command: Vec<String>) -> Result<()> {
-    let status = Command::new("podman")
+pub fn exec(conf: &Conf, name: &str, command: Vec<String>) -> Result<()> {
+    let status = Command::new(&conf.podman)
         .arg("exec")
         .arg("-it")
         .arg(name)