+++ devtools/replidev/src/podman/unshare_cmd.rs
@@ -2,12 +2,13 @@ use std::process::Command;
 
 use failure::ResultExt;
 
+use crate::Conf;
 use crate::ErrorKind;
 use crate::Result;
 
 /// Execute a command in podman unshare environment (enter the user ns but not others).
-pub fn unshare(command: Vec<&str>) -> Result<()> {
-    let status = Command::new("podman")
+pub fn unshare(conf: &Conf, command: Vec<&str>) -> Result<()> {
+    let status = Command::new(&conf.podman)
         .arg("unshare")
         .args(command)
         .status()