+++ devtools/replidev/CHANGELOG.md
@@ -6,5 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 ### Added
+- Configurable podman command to use.
 - Initialise and clean up replicante core dependencies.
+- List available `replidev deps` pods.
+- Optional per-machine configuration override.
 - Start and stop replicante core dependencies pods.